feat: ticket showing on client stand
